pyproject.toml
[project]
name = "trading-bot-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pandas-stubs==2.3.2.250926",
]
